Hero Image Added and Responsive but
I need to figure out how to get the hero img to align with the hero-text instead of the top of the container. I want both sides of the hero div to always display the same size with the top of the image aligning with the "This websi..." text and the bottom to align with the bottom of the Sign up Button.  Need to do more research.